Theme update and epithet fix
- Epithet generation will now avoid epithets of style "Era of the Mages
of the Derp Durp" and will result in "Era of the Derp Durp" instead.
- Some debug info printed in .dm above nation definitions (they are way
down. Try searching for "nation x")
- Added theme commands #racedefinition and #secondaryracedefinition that
work for example as #racedefinition "pose avvitemage". Theoretically
it's exactly the same as if you had added the line #pose avvitemage at
the end of the file. The weird exception is that very special commands -
the stuff not specifically handled in Race.java, Entity.java or
Filter.java like #chanceinc, #unitcommand and #nationcommand will look
for lines that are exactly the same except for last argument and will
replace them. The main use of this is that you can for example replace
#monsterchance or #magicpriority water instead of planning around by
removing those from the race definition and moving them forward. Be
careful though.
- Themes are generated after primary race but before secondary race
- Added themeinc "racename [not]<race> x". It will check if the
chanceinc is done to a race and if so, it will check if it is the
correct (incorrect if not switch is on, this applies to the theme
themeinc as well) race. This does not affect primary race since those
are generated before themes so that possible themes can vary racially
without huge chanceinc shenanigans.