Amazon themes, mino hand fix
- Amazon themes should be done-ish. I waffled about what #slave should
mean, and ended up going with -1 att, -1 def, -2 mor, *0.8 gcost so
they'll be cheaper but inferior, and the slavers get +1 mor, +1 gcost
(arrogant but pampered); also, some or all of the Amazon's male troops
may be slaves too. Also added a techlevel bronze, and 4 tribal themes
for each of the traditional tribes.
- Changed mino mage hand file reference; mages are still a WIP, but
it'll stop complaining about how I changed a filename