Elite (sacredgenerator based) major bugfix and tuning
The sacredgenerator elites weren't being added to nations if their poses
did not have non-elite/non-sacred roles. This caused the amount of these
units to plummet massively from the 20% that was default. In addition I
buffed the power of these slightly since they're supposed to be more
than mundane elite units.