Rewritten roster generation - no use of troopgenerator
RosterGenerator and UnitGenerator have taken whatever functionality they
had to take from TroopGenerator and are able to generate passable
rosters. Do comment.

Noteworthy stuff:
- Muuch mages had generateitem specified the wrong way and it crashed
the program. Might be that mages did not get generateitem handled
before, but I doubt it.
- Machakans had a bug with lances not being properly generated (only as
bonus weapon)
- Some new theme  chanceincs related to items
- Unit templates are now in /misc/ and the ones in /templates/ are
unused since they need to be ported. In this version 1-3 units (and
units derived from them with #maxvarieties limitations) get a special
filter at generation.
- Related to this theming things "scout", "assassin", "spy", "irregular"
may be a great idea.
- Bonus weapons are pretty rare. Some races/poses/filters should get
#bonusweaponchance <something>, but I believe I'm gonna add more nation
level randomness to it like there was in the old system so we'll see.
There is also #bonusweaponchancebonus. Non-bonus has highest value taken
and bonuschances stack on top of that. Default is 0.15 (which is not a
percentage but a value compared to a fairly complex formula)
- Dual wielding chance is 5% (given that unit has onehander and
available offhand weapons). Similarly to bonusweapons there are
#dwchance and #dwchancebonus. This is a straight percentage.

Next up: Mages and/or sacreds. Mages are going to require considerable
amount of code at least. Also probably fixing the roster/unitgen since
it's bound to be broken. I generated a few 100 nation batches without
crashes before committing, but I'm more afraid of things just being
plain broken.

It's simpler than the old loop of death in troopgenerator, but some of
the stuff in troopgenerator really was good at avoiding problems by
accident.