MonsterGenerator update
Monsters are now shapechangeunits instead of units and the monster
definitions are shapeshifts instead of poses. This allows sprite copying
and potentially some other stuff works better as well.

The avvite elephants were enabled. They had to be changed since #command
"#recolormask blablaba" just makes no sense when #command is meant to
give .dm writable stuff to units. #recolormask for them is a tag now and
it now works for secondshapes in general!

Closes #69