Dualwield and Bonusweapon chance harmonization
#dwchance
#dwchancebonus
Default dwchance for troops is 0.05 and for sacreds 0.15. Highest
#dwchance out of filters, pose, theme, race will be used and all
#dwchancebonus will be added on top of that. This is a direct percentual
chance _if_ unit has a dw-eligible mainhand and can get dw-eligible
offhand.

#bonusweaponchance
#bonusweaponchancebonus
These work the same way as before, but for troops this is just a value
in a complex algorithm. Base is random number from 0 to 1. For sacreds
base is 0.4 and it is a percentual chance, but every point of protection
from armor will also reduce the chance by 0.02

Main hand items can now be tagged by #ignore_dw_restrictions in which
case a weapon offhand can be given even if the weapon has length of over
2 or is twohanded. Offhand weapons do not need to be tagged - they are
chosen without restrictions. If you want to have two-handed offhands
wielded only with two-handed mainhands, there's #chanceinc slottag so
you can just tag mainhands with #twohander and put #chanceinc slottag
not weapon twohander *0 to the offhands.

If I can be arsed at some point, I will

Oh, and #always_dw doesn't work anymore. If it was anywhere it should be
replaced with #dwchance 1 (or  actually #dwchance 100 or something or
otherwise filters can reduce it lower with negative #dwchancebonus - not
that there are such filters right now)

I tested this by making human flails with #ignore_dw_restrictions and
now, after a fix to unit cleanup and a dumb typo fix, this closes #126