Added #givetoall and fixed/added some research related mage/priest filters
#givetoall on a mage/priest filter gives the filter to all tier 1-3
mages the filter doesn't have #notfortier for.

This is for drain immune research right now since I figured it should
appear for everyone if it appears for anyone. Especially since it is
priced at power 3. That said I made a separate hero/extramage version at
power 1 since it isn't a very great ability for those (or it can be for
extramage if it isn't cap only I suppose)