Magefilter structural changes
Default filters are now in three files:
- Strong defence
- Shapeshifts
- Rest

If you want to customize one, you need to #magefilters the two others.

On top of this I made mage and priest filters actually take into account
the mages of that tier instead of the highest tier mage for filter
selection.

Combined the results are:
- Removal of magefilters_no_shapechange and magefilters_fomorian
- Instead of whole Fomorian race, the Giants don't get shapeshifts and
have to give up a bit more for strongest defensive features.