Imperial hoburg gladiators
Well, sorta - ATM they have full resources costs. I've asked on the
forums in hopes of finding a way to set #rcost to 1 short of calculating
the size-adjusted rescost with equipment, and then using an
adustmentcommand to set #rcost to that times -1 (or more to the point,
making a rcost-specific handler for adjustment commands to do the
above).

Also, some minor bugfixes with clockworks and *their* resource costs