#fixedrescost and #adjustmentcommand move to getCommands()
Obviously this is less efficient as it will be ran quite a few times
potentially, but it is a bit stupid if something is getting resource
cost of 1 or morale of 50 and the generation has no idea it is
happening.

I tested this with hoburgs again and it managed to write some random
made up test commands, #mor 50 and #rcost so that all units cost 1
resource in game.