Gladiator & clockwork tweaks
Hoburg gladiators are almost done. The tag #cannot_be_pd needs support
added in Nation.java; I spent around an hour trying and failing to
implement the presumably-one-line change it needs in getMilitia(); if
you could take the 5 minutes it'll probably take you to do it in a
manner that's far more intuitive than what I tried, that would be
greatly appreciated.

Nerfed big clockworks a little more; -1 reinvig and mapmove (they're
supposed to be ponderous)