Ichtyid mage scaled stupidity
I agree that the prior change was probably too slavish to canon w/o
regard for balance. The following is what I scaled it back to, though
I'm not wedded to this either.

T1: +1 MR, -2 res
T2: +2 MR, -1 res
T3: +3 MR, -0 res

It may make it so it's inefficient to have hordes of T1 labrats, but
it's not the worst thing in the world if some nations are better off
with fewer stronger researchers.