determineSpecialLeadership wasn't catching all special types
* magic/demon/undead units who were e.g. getting their status from an
item (like basesprite) were not being properly processed because
commands[] hadn't had the command added at the point where dSL() was
invoked. A call to dSL() was added for all casters at the end of
polish(), although that may well not be a good place to add it - also,
it looks like non-caster heroes aren't getting special leadership added,
so it might be better if this was done for all non-scout leaders instead
of just casters.