Caelian ice gear
This should be it for 0.6.8 - although I saw a random crash that
concerns me, but am putting off locating until tomorrow.  After said
debugging, this should be ready to release.

Caelians aren't quite as cohesive as I'd intended, but they're pretty
cohesive, and that's good enough for me - the remaining variance feels
NationGeny. My biggest disappointment is that I'd tried to use
#chanceinc personalcommand below #coldres 5 *0 to ensure that only
cold-tolerant units use ice gear, but even though it's really weird to
see the rare daeva, Raptor, or Wingless in full ice kit, it's unusual
enough for me to be okay with it - although it does concern me that the
chanceinc isn't behaving like I naively assumed it would.