Slow militia
* Added a "slow militia" template that is exactly like the current
militia template, but with mapmove 1. This template is mutex with the
normal template, and will only apply to non-flying units that have an
item (base) with a #slow_militia tag (the normal militia template
selects for a lack of this tag). #slow_miltia tags were added to the
basesprite item for (non-austral; technically, these should be MM2, but
I didn't want to add a third militia version) humans of all subtypes,
(non-austral) hoburgs, herbivore lizards, humanbred Abysians (normal
Abysians have tags excluding them from the militia template), and
ichtyids; this is in keeping with vanilla militia mapmoves, except for
ichtyids, which is supposition.
* Added #cannot_be_pd tags to Warbred humanbred Abysians and heavy
clockwork hoburgs.
* Added #cannot_be_miltia/#cannot_be_pd tags to a number of poses which
are explicitly defined as elite poses, and which have pd-eligible
"lesser" variants. This was probably done more exhaustively than
necessary, as I tagged several elite/sacred-only tags as such and they
shouldn't be eligible for said roles strictly for hard-coded reasons.
Note that while such limitations were applied to "highmen" such as
Burning Ones and colossi, they were not applied to big Atlantians or
bandar, in keeping with vanilla.
* Between the above and previous #noble/#cannot_be_militia tagging, this
should effectively divide all units into slow/normal militia categories
in keeping with vanilla.
* Added sprites for advanced oriental armor (in the style of Japanese
nanban plate armor - oriental huburgs already have this); there are
plate hauberks (which are technically strictly more resource-expensive
versions of heavy samurai armor, but if one assumes that such armor is
available along class lines, this would be a more expensive upgrade to
regular samurai armor for slightly-lower-class warriors) and full plate.