Roster generation tweaks
Ability to set maximum and minimum troop counts. Consider using
minsecondaryracetroops and maxprimaryracetroops and their share
alternatives instead and/or in combination.

Ability to set a minimum to secondary affinity.

Ability to set the probability of no secondary race troops.

Also a tweak to make it possible to make small primary race only
rosters.