Dependency (#needs) rewrite and #forceslot/#forceslottype
A rewrite to how they are handled. This adds #forceslot and
#forceslottype which should guarantee slots, but I would like to warn
that these have the potential of skewing generation if used on essential
slots, especially weapons, offhands, armor and helmets.

For example, if the unit is generated a twohander and #forceslottype
gives it an onehander later, it would not have an offhand of any type
and it would be unlikely to get a bonus weapon. This can be handled by
more #forceslottype rolls, but even resource costs and armor levels of a
nation affect some rolls during generation in general, so it is not
adviseable.

The new #addthemeinc should be preferred for essential slots combined
with thisitemslottag or thisitemslottheme.

Also some documentation on #needs commands.