Foulspawn scouts
* Distinct scout pose for FS; body/limb/head types are strongly weighted
(or outright limited) to vaguely humanoid shapes (e.g. no crab heads,
but merely a reduced chance of a tentacle)
* This pose kinda looks like it's having a variant of the hat problem;
e.g. -749200513 (Ommer) has an assassin with a "tentacloid" head that
does #addthemeinc ... not large *0, but is generating with helmet "hood"
(#small/#canid) instead of "hood_large" (#large) - this may be
different, though, because I have a vague nagging suspicion that I'd
previously questioned whether scouts properly process themeincs (in
which case it may be considering nothing but the scoutgen boost for
#theme "stealthy", which both of those helmets have)