Rough montag test for FS nations
* Basic, poorly-structured montag setup for FS nations with handmade
placeholder template graphics - this is basically just the minimum I
could do to get a playable random FS nation.
* I'm seeing the elite helmet problem with these. E.g. Modmenbard
(-1809058332) unit 3034 - the Ragha fullhelmed unit with two tentacles
towards the end of the second row. Interestingly, its role is listed in
the .dm as scout even though it's not a commander. There are several
other wrong-helmet-on-a-troop-scout units in this nation as well.
* Interestingly, this nation used a template unit as a commander. It had
no special abilities, no inspiration, generic okayleadership, and the
same name as its troops, but it did allow for recruiting a random
commander.
* There are mantis-headed mindless FS (not montag units, but some of the
fixed units that were also generated) in this nation who appear to be
losing morale from a militia template despite their adjustmentcommand
"#mor 50" - this results in non-mindless units with a weird fixed 49
morale.
* Added a null-pose-name check in the montag handling function to make
it slightly more robust (previously it would crash if any pose for the
race had no name defined, even if it wasn't a montag pose)