#eliteversion fixed
In commanders only first #eliteversion was considered and was
deterministically chosen. Now multiple lines work and one is chosen
randomly if feasible. Same issue was in "get some elite helmet with same
id", in which case just whatever helmet of the same id was chosen.

Mages worked before, but using ItemSet.getRandom() was a bit dodgy when
unit chanceincs were relevant.