Foul spawn adorable wing overlay fix
Overlays were ignoring offsets. This is because they were originally
used with mounts, which used offsets to shift all other items and that
wasn't fine for overlays.

Now if an overlay has offsets defined, those are used and mount offsets
are used for overlays as well.

Protip: If you need mount overlays AND some other overlays, define the
mount ones as overlays and the other ones as overlay2 or something.