Foulspawn archers, starchild mages pt. 1
* Full (albeit narrow) implementation of Foulspawn archers - they can
either be traditional archers, spitters, rare mentalists, or throwers;
this variation is assumed to provide some unpredictability as far as
quality goes
* Foulspawn language themes - these are mandatory, but only control
nation names, unit nametypes, and in some cases, homerealm weights.
However, with 16 different themes, they should at least have
unpredictable names.
* About half of the work to implement starchild mage dynamic robes. This
was/is a different sort of headache than the above Foulspawn, so it
represents "taking a break" from that, and vice versa (the Foulspawn are
definitely the bigger headache. I can't imagine those tangled messes of
data files are anything but horrendous to figure out from a layman's
perspective.)