Foulspawn casters
* Completed basic data definitions for fixed-form FS casters. Note that
ATM none of them have hats, which is not ideal, to put it mildly.
However, as a first draft, they should be functional and the various
heads should still lead to a fair amount of variety
* Converted four of the dynastic caster masks into animal heads for FS
casters: hawk, heron, ibis, and jackal. Also added condor heads lifted
directly from the Nazca units, which admittedly kinda look like chicken
heads. I'm okay with that.
* While Foulspawn really need monsters to be *proper* Foulspawn, what's
done right now is honestly enough for them to be content-complete enough
for a release. I'm going to try to resist the siren's call of "adding
new stuff" and make an effort to get the new content to a place where
that side of the house can be locked until after the next release, so as
to maybe actually *make* a release one of these days...
* That'd probably need to consist of doing centaur mages and illithid
hybrids (basically limited versions of FS) at this point. There's a
bunch of other junk I vaguely "want" to do (FS monsters, ogres,
pureblood illithids, etc.), but realistically, if centaurs (even w/o
centaurides), FS, and hybrid-illithids get finished, that's more than
enough content for an eventual release.