NationGen garbage collection optimization pass
Supposedly moving variable declarations out of loops is best practice
for garbage collection. While there are a lot of variables for which
doing this may be helpful, simply moving Nation n outside the generation
loop in NationGen.generate() is probably the biggest single thing that
may help if this is true.