Cyprinids as a ichtyid theme
* 1/3 of ichtyid nations will be cyprinids; i.e., carpmen. This means
they'll have swamp survival and won't use coral or sharkskin gear, and
will eat twice as much as ichtyids. Tragically, Grab and Swallow has a
fixed size limit of 2 rather than being based on the wielder's size, and
while it would be hilarious to have gluttonous fishmen eating things as
big as they are, it seems a bit much. However... by giving them #trample
and #trampswallow, the desired effect can be achieved...
* Now I'm really tempted to add ichtyid-shaped samurai armor and make a
koi theme as well. I blame you for this.