SpellGen code and basic files
This is what I wrote in #518 pretty much exactly. It seems to work fine
and is enabled now, but it seriously needs more spells boringly plucked
from database before it can be enabled on a release (nation.java
contains the code that calls generator).

I'll look into making content tomorrow.