Agarthan lobos, Agarthan stat adjustment
* Pale One lobos (non-trampling); I upped their base cost for lobo
purposes to 14 because 20-hp need-not-eat, 100% DV, Str-15 lobos seemed
a bit much, even at 2/square and Att 7 (well, 6 since they use Fists
instead of Claws). I assume they'll still be worth using at a net cost
of 7g - indeed, they may still be too good as giant bags of hp even if
they're not-particularly-lethal.
* Adjusted Pale One hp down from their current NG standard of 22 to the
basegame standard of 18. Hopefully this is not an unreasonable nerf in
the wake of their recent vanilla cost reduction to 9g
* Made the lobo template grant absolute rather than relative str/att/def
modifiers; in light of this I reduced amazon thralls to 9/10/10 rather
than 9/11/11, but also reduced their cost to the standard 10g