Major bug: Foul spawn centaur graphics were borked
As long as they had a mount, the mount's offsets were used to offset the
whole rest of the image. When they became infantry, they were drawn as
infantry with random mount on the background, as normally only overlays
and mounts are able to force all other slots to use their offsets.

I added a command to force the program to use an another slot for these
purposes.