Winged Olm Commander bug fix and general CommanderGenerator tweaks
The bug was (probably) caused by the fact that the unit got commander
equipment updates before filters were copied. This was fine back in the
day, but now that filters affect what is generated it's pretty bad.

Other stuff:
- CommanderGenerator was using its own unit.getCopy(). Now it doesn't.
- Unit.getCopy() needs to copy slot memory.
- Commander generator should prioritize making commanders out of units
that cover as many needed tags as possible first instead of any unit
that covers at least one.
- Commander generator should scrap sacreds from possible commanders once
all needed commands are filled (#holy can be one of them so...). Foul
spawn nations were getting tons of sacred commanders occasionally due to
this.