Foulspawn slot weirdness
Let me preface this by admitting that my changes probably make this more
involved and less intuitive than how it was before.

This is a messy fix which works for how the files are written but may
not work with how they were supposed to be written. However, as I
couldn't see how to make them work with what had been in place, I went
ahead and "fixed" things. It also seemed like it had some redundancy in
it to start with, which made me kinda leery when changing things. The
code feels sloppy (yes, I know, I always say this), but it appears to
work fine in testing.

Basically, FS were getting *a lot* of hands and/or misc slots. Oh, and
not enough heads if they had two heads, IIRC. The old way did not allow
for poses, etc. to set a base number of slots. I think. Maybe. Some
units looked completely right and others were overblessed with limbs; in
retrospect it may be that the only change needed was commenting out what
is now line 279. The current way takes tags from the unit and uses any
"baseitemslot foo X" found as absolute basepoints in place of the 2
hands/1 head/1 body/1 feet/2 misc defaults.  All itemslot values
(positive or negative) in items are added to the baseitemslot or default
value. If no itemslot values are present, this means baseitemslot will
have defined the slot value. In all cases, the hard-coded absolute min
and max values in getItemSlots() will continue to ensure legal values.

Oh, and heads had been limited to 1; that's now increased to 2.

* I updated the itemslots.txt documentation to reflect the new mess I
made.
* Foulspawn and illithid hybrids were updated to use baseitemslot and to
dynamically reduce feet by -1 if they have crab/scorp legs or tentacles
(this will be cleaned up to 0 in the method if both their feet are
weird)
* Also, randomly, olmspawn apparently had #itemslot <value> instead of
#itemslots <value>, so that got fixed
* I found some graphical weirdness while examining this (probably
because of me abusing #needs), but that's less important than this