Command chanceinc "fix"
Originally command related chanceincs were supposed to have the argument
commands with the # in front of them. It's not like even I remembered
this so no wonder this is causing issues.

I made all (?) command related chanceincs accept the commands with our
without # in front of them as it should hardly cause any issues. I
generally recommend using the commands with # with everything, but for
chanceincs themselves this should not be necessary now.