Some sacredgen tweaks
There were a few bugs and I tuned some stuff a bit. This closes #761 for
now. I did write a large share of a new system but in the end the main
differences of it and the current system would've been how stat boosts /
magic weapon / etc are distributed and I didn't feel like rethinking all
that.

Magic weapon granting could probably use a small adjustment though to
fetch the weapons instantly instead of assigning a vague "use this much
power when equipping", but that's a really tiny thing.