Grand CustomItemGen fix
This closes #786

Basically CustomItemGen handled fetching item data from weapon database
and converting things to Dominions syntax. Part of this was fixing
command names (#dt_blunt -> #blunt) and part was handling boolean
variables (a non-blunt weapon shouldn't print at all instead of #blunt 0
and a blunt weapon should print as #blunt instead of #blunt 1).  This
was entirely manually handled and included a pile of variables that did
not need either operations but which items might need.

Now there's automation. It checks from a big list of if-statements
whether an attribute in database should be skipped on inputted in a
special manner and if not, it'll write it as a boolean (don't write
argument, write anything only if value in database is 1) or as a
non-boolean (write everything) based on the database. Database decides
that if a variable is ever only empty, 0 or 1, it's a boolean. I can't
think of any case where this would cause issues.