Removed unused import java.util.function.Consumer. Cleaned up some whitespace and a couple unused variables/unused variable assignments. Converted "x.size() == 0" to "x.isEmpty()" in several places. Not a significant difference in speed, but easier to read.
Finished up converting all the List declarations to diamond inference. For example:
"List<Race> items = new ArrayList<>()" instead of "List<Race> items = new ArrayList<Race>()", as the second <Race> is unnecessary.

Removed another exception catching for parsing integers out of streams and replaced it with a simple boolean check. This should speed things up slightly.

Renamed the variable units in handleShapeShifts() to avoid shadowing the variable units declared in the begininng of NationGen.java.

Converted a group of if statements in handleShapeShift() to a switch for easier readability.