Hoburg clockwork special stuff fix via #additionalweaponslot
Main issue was that #generateitem happens very early these days. It used
to be done later but it needs to be done at this point to do what it
should do for troops: Make all the weapon variants match.

I cannot remember #generateitem it was done later originally but then
changed, but oh well, there's a potential problem in using this sort of
approach anyway: You could get two identical variants of a clockwork
unit with a fire or shock lance.

The new command generates late and uses similar restrictions on items as
weapons to make consistent rosters:

#additionalweaponslot <probability> <slot>
As above, but with three differences:
1) This is done late, so it can replace weapons and such without issues
2) It works only for troops, ie troops, sacreds, elites and such
3) It treats variety like for weapons: If you have 3 units with the same
exact armor, the item from this command will vary if possible and try
to primarily use items from other armor levels of the same pose if such
exist.

For mages and priests #forceslot would be better. That would actually
work here if it was added to the flame effect to force the weapon to be
lance, but eh, no.

Extra note: This can be used for multi-arm units in troop rosters. If
there's need to really implement multi-arm units it'd be possible to
extend this to all units pretty easily. After all, theoretically
#additionalweaponslot 1 extra_arm would generate an item from extra arm
in a way that would make the weapon vary for each unit variant and
#forceslot can be used to force the second extra arm of a four armed
unit to be null when the first one carries a twohander. I'm not entirely
sure it needs to even vary though as the two primary hands would
already.