Oriental human/tengu armor tweaks
* Gave Jomon-style armor Jomon-style legs so their legs don't look quite
as lumpy
* Added integrated armored footware for a couple of armors
* Fixed some wierdness with missing bits of a few armors - a number of
sprites were missing 1-8 pixels around the edges of the suit of armor
* Did some minor graphic tweaking of a couple of things
* Ended up re-orienting all oriental gear sprites to be placed in
reference to humans rather than tengu; this entailed a bunch of changes
(including some for medium clockworks) but the data files are a lot
cleaner (and the sprite placement is in reference to the most common
nation type to use them). If I'm feeling brave at some point, I may do a
careful multi-file replaceall to rename /graphics/tengu/ to
/graphics/oriental/, but not today...
* Noticed that mounted samurai armor is wierd - it's one-handed troops
using mounted armor sprites based on two-handed infantry armor. For the
time being, I did not change this, as it looks mostly okay with the
"normal human" shirts that cavalry all have providing sleeves for them.
This is pretty plainly based on these poses originally not using any TC
gear and just having no-dachis and katanas, but since that's no longer
the case I may circle back and change this eventually
* I also noticed that I'd made sprites for advanced oriental armor of
the same sort that oriental hoburgs have - basically Jomonese plate
hauberks and full plate based on Japanese nanban armor - so I added them
in with low probability only when the techlevel is advanced