Remove mapmove adjustments from code
Dominions 5 reduces mapmove based on armor already. MM2 seems to become mapmove of 14. Manual adjustments to quite a few poses/races should be done to capture some of the more subtle variations in game (Machakan militia is base mm 12 while normal troops are mm 16 for example), but for now this shall suffice.