Chanceinc Logic Operators
All the existing files should still work fine.

Testing has been limited in practice, but I did test the logical operation parser quite a bit before I plugged it in.

I removed the call handleChanceincs for a list of units. This made very little sense for anything except mage naming and possibly equipping since it's a bit ludicrous to, for example, reduce the probability of something to half 3 times because you have 3 lizard mages or something. Mage naming at least has to move to use themeincs sooner or later, but it seems to work fine enough until a bigger rework now that I just made all the List<Unit> calls use .get(0) as the call. The real reason for the removal was mostly that I couldn't figure a sensible way to allow combining unit and nation chanceincs in logic operations without this, but hey, nothing of value was lost.

Documentation is included.