MageNamer fix
It began crashing with my list.get(0) calls I guess. These failsafes make it work like it used to and are probably slightly more efficient too.

Mage naming should use themeincs to an extent, but no point in improving it unless we're doing a major rework while we're at it.