Some more flag icons
Ported some low-hanging-fruit as far as flag icons go. The serpent was
literally the small serpent icon resized, so I did the same with the
bone serpent and tentacle. Ur's ziqurrat was also just scaled up, so I
did likewise with the pyramid, though I did smooth that one.  The
plumage icons were based on fairly large pretender graphics, so they
work all right even without scaling up; same with the horror. The large
number of plumage icons is based on an assumption that there is no color
masking.