Wall defenders priority to #castledef > 0
* Very crude, heavy weighting towards having castle defense specialists
chosen out of whatever list of potential defenders was picked, as this
is more or less their raison d'être
* There should probably include some weighting against #mindless,
#undisciplined, and #animal units, and for #flying; all things being
equal, units that are on the walls should be units that are good at
siege defense since they contribute to a fort's siege defense value
* This reminds me - monkeys are not supposed to get the animal siege
defense penalty; I'm pretty sure ours will, though. I'll check and add
an issue to figure out the best fix if they are.