Fix the very broken new chanceinc handling
Now not <statement> should mean the opposite of <statement> and any nots inside the statement should be considered statement by statement individually.

Something might still be broken because I was misguided when rewriting the chanceincs themselves for the new system, but the broken ones should be a tiny minority. If something seems super weird regarding choice of filters or equipment, do mention.