Even more ChanceIncHandler rewriting!
Now pretty much everything should accept both:

not <chanceinc> <args> value

and

<chanceinc> not <args> value

For some chanceincs there is a difference. These are generally multi-argument chanceincs. For example, all the "slot has tag" ones tend to work better with the latter example, since then it is true if slot x contains said theme and false otherwise. Former example would instead be true as long as slot x did not contain the tag, which is kind of disastrous for stuff like "this unit should not get any chest armor which isn't for armored units" since it turns it to pretty much the opposite. This was the issue before with the particular foul spawn, but it was fixed separately.