Non-intristic non-str based range ranged weapons with #nostr get #bowstr instead in database conversion
There are definitely going to be issues with this at some point, but I cannot think of any missile weapons generated to sacreds and elites from base game items that shouldn't have #bowstr but has a defined range and isn't a bonus weapon.

Illwinter should make the determination of #bowstr obvious in data, which it isn't. The place where #nostr is defined still indeed defines bows with it and I cannot find anything meaningful separating a bow and a spear.