Filter fix
Removed/simplified conditional chanceincs from mage filters, as there
appears to be a problem with them, or possibly just with personalmagic
and them. I'm not alert enough tonight to try to figure out if the
problem is conditionals, personalmagic, or my use of conditionals, but I
may revisit examining this when I'm more able.