Fix conditional chanceincs (?)
The main issue was that for some reason there was no .toLowerCase() anywhere and thus "OR" and "AND" weren't considered operators.

There was also a crash bug related to chanceinc handling returning null when it doesn't hit a suitable chanceinc at all. This is correct behavior to separate it from failing a chanceinc, but it was handled poorly.