Dantih, done
* Robes for dantih casters
* A bunch of stupid bugs fixed based based on generating a couple
hundred nations
* A whole bunch of graphics tweaks based on generating a couple hundred
nations
* I'm not 100% happy with how they look - e.g. the falchion pretty much
touching their head looks off to me, but not enough to prompt me to
change it - but I'm *fairly* happy - and when it works, it works pretty
well IMO
* I still could use to add demographic themes so that mammoth men are
possible, but this is good enough for now. Likewise Austral regional
themes for African-elephantmen, or if I were ambitious enough to do all
the spriting, elephentmen in samurai armor, hehe.
* Definitions pass SpriteChecker, but I've not done in-game testing
* Probably completely unbalanced