Fixed magepriest generation for extra mages
There is code to give extra mages a 20% chance to have priest paths if the main mage line has none, but the actual generation is gated behind magePriests == true. But extra mages are never set as magepriests unless magePriests == false. This forces magepriests = true when compensationMagePriests is set for extra mages. (Standard magePriests are still not generated unless compensationMagePriests == false, so this does not change existing magepriest functionality.)

This does have some balance implications since it gives an extra shot at non-slow-to-recruit H3s that have valuable magic paths, so I added a check to force H3 extra mages to always be slow to recruit.