Clean up and consolidate leadership calculation for mages/priests
New method randomizeLeadership replaces largely redundant if/else trees in generateMages and generatePriests. There are some minor nuances of the old generation that are not captured but for the most part it's a fairly accurate approximation of what the old method would have generated, hugely simplified and condensed. Could probably be expanded to encompass mundane commanders.