Cleaned up and condensed extra mage chances
Consolidated a lot of conditionals for extra mage chances and removed a redundant boolean. Preserved existing logic for all extra mage chances, except as noted there was an extra 1 in 10 for nations without guaranteed earth 1/fire 1 only - this seemed very arbitrary for such a small change that I just took out the restriction.