Nadir "Fix"
Basically, the Nadir were even more broken than usually assumed. The following is my reasoning for the chance values, tried to follow patterns, but this will likely make them more imbalanced due to Iron being a higher chance.Anyway:

Primitive: The numbers kept going down as age increased, so I followed that.
Bronze: These dudes are the resurrected corpses of agarthans. Given that waxing and zenith are what I associate with the age of the Cavernwights, I assumed this would be the most common theme as they just use ancestral weapons and armor that refuses to rust.
Iron: On the other hand, the peeps who ressurrected the Cavernwights probably had access to ironworking and the Cavernwights picked up on that.

Funny story, I was trying to generate primitive LA agarthans and I kept getting Nadir. That just didn't do, because I wanted the weakest race possible. So I opened up the .dm and found out this abomination.